+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - Fix sticky footer bug in IE11.
   [#347](https://github.com/OSC/ood-dashboard/issues/347)
+- Bump version of gem dependencies for security fixes.
 
 ## [1.23.0] - 2018-02-26
 ### Added