+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- (Batch Connect) Can now configure whether global configuration is used for
+  system apps.
 
 ## [1.24.0] - 2018-03-27
 ### Changed